# server/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Gram API Description
    description: Gram is the tools platform for AI agents
    version: 0.0.1
servers:
    - url: http://localhost:80
      description: Default server for gram
paths:
...
    /rpc/toolsets.create:
        post:
            tags:
                - toolsets
            summary: createToolset toolsets
            description: Create a new toolset with associated tools
            operationId: toolsets#createToolset
            parameters:
                - name: X-Gram-Session
                  in: header
                  description: Session header
                  allowEmptyValue: true
                  schema:
                    type: string
                    description: Session header
                    example: Ea voluptatem veniam accusantium consectetur eaque.
                  example: Tenetur et rerum.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateToolsetForm'
                        example:
                            description: Officia voluptas aspernatur quibusdam.
                            http_tool_ids:
                                - Consequatur itaque.
                                - Modi aliquam aut provident incidunt suscipit.
                            name: Rerum animi expedita.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Toolset'
                            example:
                                created_at: Tenetur sit accusantium.
                                description: Error et enim nostrum doloremque.
                                http_tool_ids:
                                    - Error voluptatum dolore debitis eos doloribus magnam.
                                    - Aut sunt.
                                    - Occaecati assumenda ut consectetur et ut velit.
                                    - Veniam et.
                                id: Consequatur deserunt sequi maxime.
                                name: Fuga id ea et esse.
                                organization_id: Doloribus ut quaerat aut.
                                project_id: Assumenda quas dolor mollitia laboriosam voluptatem.
                                updated_at: Vel sit consequuntur est.
            security:
                - gram_session_header_X-Gram-Session: []
    /rpc/toolsets.get/{id}:
        get:
            tags:
                - toolsets
            summary: getToolsetDetails toolsets
            description: Get detailed information about a toolset including full HTTP tool definitions
            operationId: toolsets#getToolsetDetails
            parameters:
                - name: id
                  in: path
                  description: The ID of the toolset
                  required: true
                  schema:
                    type: string
                    description: The ID of the toolset
                    example: Eum ex necessitatibus.
                  example: Nostrum labore.
                - name: X-Gram-Session
                  in: header
                  description: Session header
                  allowEmptyValue: true
                  schema:
                    type: string
                    description: Session header
                    example: Eaque perspiciatis fuga excepturi alias.
                  example: Expedita quod vero aperiam.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ToolsetDetails'
                            example:
                                created_at: Illo occaecati non vel magni.
                                description: Rerum debitis culpa molestiae.
                                http_tools:
                                    - apikey_env_var: Et aut consequatur.
                                      bearer_env_var: Et aut consequatur provident animi.
                                      body_schema: Quis unde sit rerum.
                                      description: Quia deserunt et at et.
                                      headers_schema: Quis enim error labore vero ut.
                                      http_method: Ullam quod occaecati consectetur voluptatem animi aut.
                                      id: Architecto odit quo.
                                      name: Deserunt tenetur.
                                      password_env_var: Eos dignissimos suscipit architecto culpa praesentium.
                                      path: Non temporibus at expedita magni perferendis amet.
                                      pathparams_schema: Repellendus vel sed.
                                      queries_schema: Iure perferendis et deleniti quia.
                                      security_type: Aut tenetur.
                                      server_env_var: Laudantium perferendis dolorum ut rem a.
                                      username_env_var: Accusantium occaecati eum in voluptatum nemo mollitia.
                                    - apikey_env_var: Et aut consequatur.
                                      bearer_env_var: Et aut consequatur provident animi.
                                      body_schema: Quis unde sit rerum.
                                      description: Quia deserunt et at et.
                                      headers_schema: Quis enim error labore vero ut.
                                      http_method: Ullam quod occaecati consectetur voluptatem animi aut.
                                      id: Architecto odit quo.
                                      name: Deserunt tenetur.
                                      password_env_var: Eos dignissimos suscipit architecto culpa praesentium.
                                      path: Non temporibus at expedita magni perferendis amet.
                                      pathparams_schema: Repellendus vel sed.
                                      queries_schema: Iure perferendis et deleniti quia.
                                      security_type: Aut tenetur.
                                      server_env_var: Laudantium perferendis dolorum ut rem a.
                                      username_env_var: Accusantium occaecati eum in voluptatum nemo mollitia.
                                    - apikey_env_var: Et aut consequatur.
                                      bearer_env_var: Et aut consequatur provident animi.
                                      body_schema: Quis unde sit rerum.
                                      description: Quia deserunt et at et.
                                      headers_schema: Quis enim error labore vero ut.
                                      http_method: Ullam quod occaecati consectetur voluptatem animi aut.
                                      id: Architecto odit quo.
                                      name: Deserunt tenetur.
                                      password_env_var: Eos dignissimos suscipit architecto culpa praesentium.
                                      path: Non temporibus at expedita magni perferendis amet.
                                      pathparams_schema: Repellendus vel sed.
                                      queries_schema: Iure perferendis et deleniti quia.
                                      security_type: Aut tenetur.
                                      server_env_var: Laudantium perferendis dolorum ut rem a.
                                      username_env_var: Accusantium occaecati eum in voluptatum nemo mollitia.
                                    - apikey_env_var: Et aut consequatur.
                                      bearer_env_var: Et aut consequatur provident animi.
                                      body_schema: Quis unde sit rerum.
                                      description: Quia deserunt et at et.
                                      headers_schema: Quis enim error labore vero ut.
                                      http_method: Ullam quod occaecati consectetur voluptatem animi aut.
                                      id: Architecto odit quo.
                                      name: Deserunt tenetur.
                                      password_env_var: Eos dignissimos suscipit architecto culpa praesentium.
                                      path: Non temporibus at expedita magni perferendis amet.
                                      pathparams_schema: Repellendus vel sed.
                                      queries_schema: Iure perferendis et deleniti quia.
                                      security_type: Aut tenetur.
                                      server_env_var: Laudantium perferendis dolorum ut rem a.
                                      username_env_var: Accusantium occaecati eum in voluptatum nemo mollitia.
                                id: Perspiciatis nostrum minus repellendus dolore qui.
                                name: Nobis enim quas dolorem et inventore.
                                organization_id: Sapiente aliquam sit aperiam.
                                project_id: Unde sint nemo rerum recusandae harum.
                                updated_at: Modi suscipit laboriosam voluptatum voluptates dolore.
            security:
                - gram_session_header_X-Gram-Session: []
    /rpc/toolsets.list:
        get:
            tags:
                - toolsets
            summary: listToolsets toolsets
            description: List all toolsets for a project
            operationId: toolsets#listToolsets
            parameters:
                - name: X-Gram-Session
                  in: header
                  description: Session header
                  allowEmptyValue: true
                  schema:
                    type: string
                    description: Session header
                    example: Nesciunt porro officiis non.
                  example: Odio voluptas aut excepturi est.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListToolsetsResult'
                            example:
                                toolsets:
                                    - created_at: Perspiciatis sed facere sed.
                                      description: Magni et quia consectetur voluptatibus voluptas iusto.
                                      http_tool_ids:
                                        - Deleniti sint.
                                        - Consequatur dolorum autem ut totam sint.
                                      id: Cum dignissimos quis modi velit maiores.
                                      name: Reiciendis aspernatur.
                                      organization_id: Nihil mollitia.
                                      project_id: Ab dolorum repellendus aspernatur voluptatem aspernatur.
                                      updated_at: Corrupti harum voluptates autem impedit consequatur consequatur.
                                    - created_at: Perspiciatis sed facere sed.
                                      description: Magni et quia consectetur voluptatibus voluptas iusto.
                                      http_tool_ids:
                                        - Deleniti sint.
                                        - Consequatur dolorum autem ut totam sint.
                                      id: Cum dignissimos quis modi velit maiores.
                                      name: Reiciendis aspernatur.
                                      organization_id: Nihil mollitia.
                                      project_id: Ab dolorum repellendus aspernatur voluptatem aspernatur.
                                      updated_at: Corrupti harum voluptates autem impedit consequatur consequatur.
                                    - created_at: Perspiciatis sed facere sed.
                                      description: Magni et quia consectetur voluptatibus voluptas iusto.
                                      http_tool_ids:
                                        - Deleniti sint.
                                        - Consequatur dolorum autem ut totam sint.
                                      id: Cum dignissimos quis modi velit maiores.
                                      name: Reiciendis aspernatur.
                                      organization_id: Nihil mollitia.
                                      project_id: Ab dolorum repellendus aspernatur voluptatem aspernatur.
                                      updated_at: Corrupti harum voluptates autem impedit consequatur consequatur.
                                    - created_at: Perspiciatis sed facere sed.
                                      description: Magni et quia consectetur voluptatibus voluptas iusto.
                                      http_tool_ids:
                                        - Deleniti sint.
                                        - Consequatur dolorum autem ut totam sint.
                                      id: Cum dignissimos quis modi velit maiores.
                                      name: Reiciendis aspernatur.
                                      organization_id: Nihil mollitia.
                                      project_id: Ab dolorum repellendus aspernatur voluptatem aspernatur.
                                      updated_at: Corrupti harum voluptates autem impedit consequatur consequatur.
            security:
                - gram_session_header_X-Gram-Session: []
    /rpc/toolsets.update/{id}:
        post:
            tags:
                - toolsets
            summary: updateToolset toolsets
            description: Update a toolset's properties including name, description, and HTTP tools
            operationId: toolsets#updateToolset
            parameters:
                - name: id
                  in: path
                  description: The ID of the toolset to update
                  required: true
                  schema:
                    type: string
                    description: The ID of the toolset to update
                    example: Consequatur velit molestiae mollitia voluptatem consequuntur architecto.
                  example: Cumque rem velit.
                - name: X-Gram-Session
                  in: header
                  description: Session header
                  allowEmptyValue: true
                  schema:
                    type: string
                    description: Session header
                    example: Aut aut neque quisquam error.
                  example: Alias et doloremque dolorem tempore similique voluptatum.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateToolsetRequestBody'
                        example:
                            description: Quaerat illo qui et.
                            http_tool_ids_to_add:
                                - Dolores ipsum amet perspiciatis.
                                - Et dignissimos veniam vel.
                                - Iste id suscipit corrupti ea et quaerat.
                            http_tool_ids_to_remove:
                                - Molestias corrupti voluptas.
                                - Dolor nisi nesciunt et dolores.
                                - Aut ipsam voluptate non iusto et ut.
                            name: Quos quis fugit.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Toolset'
                            example:
                                created_at: Corrupti non ipsa architecto.
                                description: Praesentium voluptas dolor magnam voluptates est.
                                http_tool_ids:
                                    - Labore incidunt.
                                    - Dolor deleniti porro.
                                    - Rerum ut reiciendis nisi et totam saepe.
                                    - Blanditiis qui vel explicabo sint rem voluptatem.
                                id: Quasi enim.
                                name: Velit aspernatur qui.
                                organization_id: Non consectetur.
                                project_id: Omnis eos.
                                updated_at: Accusantium cumque.
            security:
                - gram_session_header_X-Gram-Session: []
components:
    schemas:
...
        CreateToolsetForm:
            type: object
            properties:
                description:
                    type: string
                    description: Description of the toolset
                    example: Quas laudantium inventore.
                http_tool_ids:
                    type: array
                    items:
                        type: string
                        example: Nulla repellendus.
                    description: List of HTTP tool IDs to include
                    example:
                        - Et omnis ut.
                        - Eligendi qui quibusdam non enim.
                        - Nulla omnis est.
                        - Quis dignissimos molestias quo quia dolorem.
                name:
                    type: string
                    description: The name of the toolset
                    example: Dolores incidunt non similique veniam qui eaque.
            example:
                description: Autem et.
                http_tool_ids:
                    - Dolores non.
                    - Ex facere.
                    - Temporibus veniam eos fugit nisi iusto quasi.
                    - Quia consequatur beatae nisi reprehenderit.
                name: Excepturi consequatur.
            required:
                - name
...
        ListToolsetsResult:
            type: object
            properties:
                toolsets:
                    type: array
                    items:
                        $ref: '#/components/schemas/Toolset'
                    description: The list of toolsets
                    example:
                        - created_at: Totam explicabo porro.
                          description: Repellendus explicabo.
                          http_tool_ids:
                            - Quae fugiat dicta cum nesciunt.
                            - Et est et eveniet sed.
                          id: Accusantium in ipsa deserunt non.
                          name: Maxime rem.
                          organization_id: Totam praesentium qui aut.
                          project_id: Sed veritatis aspernatur natus ipsam aut officia.
                          updated_at: Quasi delectus ab officia.
                        - created_at: Totam explicabo porro.
                          description: Repellendus explicabo.
                          http_tool_ids:
                            - Quae fugiat dicta cum nesciunt.
                            - Et est et eveniet sed.
                          id: Accusantium in ipsa deserunt non.
                          name: Maxime rem.
                          organization_id: Totam praesentium qui aut.
                          project_id: Sed veritatis aspernatur natus ipsam aut officia.
                          updated_at: Quasi delectus ab officia.
                        - created_at: Totam explicabo porro.
                          description: Repellendus explicabo.
                          http_tool_ids:
                            - Quae fugiat dicta cum nesciunt.
                            - Et est et eveniet sed.
                          id: Accusantium in ipsa deserunt non.
                          name: Maxime rem.
                          organization_id: Totam praesentium qui aut.
                          project_id: Sed veritatis aspernatur natus ipsam aut officia.
                          updated_at: Quasi delectus ab officia.
            example:
                toolsets:
                    - created_at: Totam explicabo porro.
                      description: Repellendus explicabo.
                      http_tool_ids:
                        - Quae fugiat dicta cum nesciunt.
                        - Et est et eveniet sed.
                      id: Accusantium in ipsa deserunt non.
                      name: Maxime rem.
                      organization_id: Totam praesentium qui aut.
                      project_id: Sed veritatis aspernatur natus ipsam aut officia.
                      updated_at: Quasi delectus ab officia.
                    - created_at: Totam explicabo porro.
                      description: Repellendus explicabo.
                      http_tool_ids:
                        - Quae fugiat dicta cum nesciunt.
                        - Et est et eveniet sed.
                      id: Accusantium in ipsa deserunt non.
                      name: Maxime rem.
                      organization_id: Totam praesentium qui aut.
                      project_id: Sed veritatis aspernatur natus ipsam aut officia.
                      updated_at: Quasi delectus ab officia.
            required:
                - toolsets
...
        Toolset:
            type: object
            properties:
                created_at:
                    type: string
                    description: When the toolset was created
                    example: Dolorum commodi.
                description:
                    type: string
                    description: Description of the toolset
                    example: Accusantium eaque aut alias repellendus.
                http_tool_ids:
                    type: array
                    items:
                        type: string
                        example: Possimus praesentium nam.
                    description: List of HTTP tool IDs included in this toolset
                    example:
                        - Quia et.
                        - Et consequatur molestiae dicta quidem est.
                        - Quibusdam et.
                        - Facilis occaecati aut esse recusandae ut et.
                id:
                    type: string
                    description: The ID of the toolset
                    example: Est illum.
                name:
                    type: string
                    description: The name of the toolset
                    example: Dolore quidem totam.
                organization_id:
                    type: string
                    description: The organization ID this toolset belongs to
                    example: Velit eius esse veniam molestiae magni.
                project_id:
                    type: string
                    description: The project ID this toolset belongs to
                    example: Dolores est officia ducimus doloribus temporibus consectetur.
                updated_at:
                    type: string
                    description: When the toolset was last updated
                    example: Odio eaque magni vitae odio quo unde.
            example:
                created_at: Dolorem officiis quaerat voluptatem recusandae.
                description: Sint dolores sit laboriosam et dolores rerum.
                http_tool_ids:
                    - Nihil omnis a dignissimos rerum nostrum non.
                    - Voluptas quo vitae amet voluptatum illum aut.
                id: Distinctio aliquam aut vero.
                name: Reprehenderit fugiat temporibus reprehenderit.
                organization_id: Placeat aut nam nostrum alias.
                project_id: Mollitia perspiciatis totam.
                updated_at: Similique dolor dolorum.
            required:
                - id
                - project_id
                - organization_id
                - name
                - created_at
                - updated_at
...
        UpdateToolsetRequestBody:
            type: object
            properties:
                description:
                    type: string
                    description: The new description of the toolset
                    example: Itaque non nam quisquam voluptas eveniet.
                http_tool_ids_to_add:
                    type: array
                    items:
                        type: string
                        example: Sint voluptatibus ut sequi est quos qui.
                    description: HTTP tool IDs to add to the toolset
                    example:
                        - Est pariatur eum distinctio debitis blanditiis temporibus.
                        - Vitae consectetur dolorem.
                http_tool_ids_to_remove:
                    type: array
                    items:
                        type: string
                        example: Libero aut.
                    description: HTTP tool IDs to remove from the toolset
                    example:
                        - Consequatur nesciunt.
                        - Sint et in nihil.
                        - Ut nam dolor provident corporis accusamus.
                        - Qui quis dolore.
                name:
                    type: string
                    description: The new name of the toolset
                    example: Corrupti eos quasi ad fuga voluptatibus modi.
            example:
                description: Delectus impedit dicta.
                http_tool_ids_to_add:
                    - Ut ut harum.
                    - Vitae ex aspernatur sequi.
                    - Voluptatem aut autem dolorum.
                http_tool_ids_to_remove:
                    - Sit eius quia rem.
                    - Aspernatur beatae.
                name: Quia omnis odit laboriosam.
    securitySchemes:
        gram_session_header_X-Gram-Session:
            type: apiKey
            description: Gram Session based auth. By cookie or header.
            name: X-Gram-Session
            in: header
...